requirements.txt
Werkzeug==2.2.2
Flask==2.1.3
fastapi==0.102.0
uvicorn==0.23.2
jsonschema==4.17.3
pyhumps==3.8.0
pymongo==3.12.0
Requests==2.31.0
tqdm==4.65.0